+++ apoinfo.py
@@ -20,11 +20,9 @@ __version__ = (0, 1, 25)
 # scope: inline
 # scope: hikka_only
 # scope: hikka_min 1.2.10
-# requires: emoji
 
 import logging
 
-import emoji  # skipcq: PY-W2000
 import git
 from telethon.tl.types import Message
 from telethon.utils import get_display_name
@@ -278,7 +276,7 @@ class ApodiktumInfoMod(loader.Module):
         self._db = db
         self._client = client
         self.apo_lib = await self.import_lib(
-            "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
+            "https://raw.githubusercontent.com/anon97945/hikka-libs/master/apodiktum_library.py",
             suspend_on_error=True,
         )
         self.apo_lib.apodiktum_module()