+++ apolib_controller.py
@@ -19,10 +19,8 @@ __version__ = (0, 1, 3)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
-# requires: emoji
 
 import logging
-import emoji  # skipcq: PY-W2000
 
 from telethon.tl.types import Message
 
@@ -76,7 +74,7 @@ class ApodiktumLibControllerMod(loader.Module):
         self._db = db
         self._client = client
         self.apo_lib = await self.import_lib(
-            "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
+            "https://raw.githubusercontent.com/anon97945/hikka-libs/master/apodiktum_library.py",
             suspend_on_error=True,
         )
         self.apo_lib.apodiktum_module()