+++ apolib_tester.py
@@ -19,11 +19,9 @@ __version__ = (0, 0, 12)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
-# requires: emoji
 
 import logging
 
-import emoji  # skipcq: PY-W2000
 from telethon.tl.types import Message
 
 from .. import loader, utils
@@ -65,7 +63,7 @@ class ApoLibTesterMod(loader.Module):
         self.db = db
         self.client = client
         self.apo_lib = await self.import_lib(
-            "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
+            "https://raw.githubusercontent.com/anon97945/hikka-libs/master/apodiktum_library.py",
             suspend_on_error=True,
         )
         self.apo_lib.apodiktum_module()