+++ gtranslate.py
@@ -19,11 +19,10 @@ __version__ = (0, 0, 66)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
-# requires: googletrans==4.0.0-rc1 emoji
+# requires: googletrans==4.0.0-rc1
 
 import logging
 
-import emoji  # skipcq: PY-W2000
 import googletrans
 from telethon.tl.types import Message
 
@@ -132,7 +131,7 @@ class ApodiktumGTranslateMod(loader.Module):
         self._db = db
         self._client = client
         self.apo_lib = await self.import_lib(
-            "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
+            "https://raw.githubusercontent.com/anon97945/hikka-libs/master/apodiktum_library.py",
             suspend_on_error=True,
         )
         self.apo_lib.apodiktum_module()