+++ lcr.py
@@ -19,12 +19,10 @@ __version__ = (0, 0, 30)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
-# requires: emoji
 
 import asyncio
 import logging
 
-import emoji  # skipcq: PY-W2000
 from telethon import events
 from telethon.tl.types import Message
 
@@ -111,7 +109,7 @@ class ApodiktumLCRMod(loader.Module):
         self._db = db
         self._client = client
         self.apo_lib = await self.import_lib(
-            "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
+            "https://raw.githubusercontent.com/anon97945/hikka-libs/master/apodiktum_library.py",
             suspend_on_error=True,
         )
         self.apo_lib.apodiktum_module()