+++ msg_merger.py
@@ -19,12 +19,12 @@ __version__ = (0, 0, 39)
 
 # scope: hikka_only
 # scope: hikka_min 1.2.11
-# requires: emoji
+
 
 import logging
 from datetime import datetime, timezone
 
-import emoji  # skipcq: PY-W2000
+
 from telethon.errors import MessageIdInvalidError, MessageNotModifiedError
 from telethon.tl.types import Message
 
@@ -213,7 +213,7 @@ class ApodiktumMsgMergerMod(loader.Module):
         self._db = db
         self._client = client
         self.apo_lib = await self.import_lib(
-            "https://raw.githubusercontent.com/anon97945/hikka-mods/master/apodiktum_library.py",
+            "https://raw.githubusercontent.com/anon97945/hikka-libs/master/apodiktum_library.py",
             suspend_on_error=True,
         )
         self.apo_lib.apodiktum_module()